==> 1.Preprocess_raw.ipynb <==
import os
import pandas as pd

# set working directory here
os.chdir(r'Replication_TrainingDatasetCensorship/output') # set working directory

# change (file) name to preprocess each word embedding
name = ''
# original_sgns.renmin
# original_sgns.baidubaike
# original_sgns.wiki
data = []
counter = 0
with open(name, 'r') as f:
    for line in f:
        if counter==0:
            pass
        else:
            d = line.rstrip().split()
            if len(d)!=301:
                print("incorrect length")
                if len(d)==302:
                    new = d[2:]
                    new.insert(0, d[0]+d[1])
                    if len(new)==301:
                        print(new[0:2])
                        data.append(new)
                    else:
                        print(len(new))
            else:
                data.append(d)
        counter += 1

res = pd.DataFrame(index=range(len(data)), columns=range(301))
for i in range(len(data)):
    res.loc[i] = data[i]
len(data)

res = res.dropna()
res.drop_duplicates(subset=[0], inplace=True)
res = res.reset_index(drop=True)
res.head()

# change file name to save each of the preprocessed word embedding
out = ''
# 1.tocsv_renmin_word.csv
# 1.tocsv_baidubaike_word.csv
# 1.tocsv_wiki_word.csv

res.to_csv(out, header=False, index=False)
