==> 4.News_sentiment_prediction_nbsvm.ipynb <==
import re
import os
import pandas as pd
import numpy as np
import math
import random
import jieba
import sys
from sklearn.naive_bayes import GaussianNB
from sklearn import svm
from sklearn import metrics
from sklearn import preprocessing

print(sys.version)

def word2idx(embedding):
    wordidx = {}
    for i in range(len(embedding)):
        wordidx[embedding['name'][i]] = i
    return(wordidx)
    
def sen2idx(sen, wordidx):
    idxword = []
    for s in sen:
        try:
            idx = wordidx[s]
            idxword.append(idx)
        except:
            continue
    return(idxword)

os.chdir(r'Replication_TrainingDatasetCensorship/') # set working directory

jieba.load_userdict("auxiliary/Dict.txt")

words =[# freedom
    "自由", "言论自由", "集会自由", "新闻自由", "结社自由", "自由权", "民主自由", "自由言论", "创作自由", 
    "婚姻自主", "自由民主", "自由市场", "自决", "自决权", "生而自由", "自由自在", "自由选择", "自由思想", 
    "公民自由", "自由竞争", "宗教自由", "自由价格", 
    # democracy
    "民主", "自由民主", "民主自由", "民主制度", "民主化", "社会民主主义", "民主运动", "民主主义", "民主改革", 
    "民主制", "民主选举", "民主权力", "多党制", "民主法制", "民主权利" ,
    # election
    "选举", "直接选举", "议会选举", "间接选举", "直选", "换届选举", "民选", "投票选举", "全民公决", "总统大选", 
    "大选", "普选", "全民投票", "民主选举",
    # collective action
    "抗议", "示威", "示威游行", "示威抗议", "游行示威", "静坐示威", "绝食抗议", "请愿", "示威运动", "游行", 
    "罢教", "静坐", "集会游行", "罢课", "签名运动", 
    # negative figures
    "林彪", "王洪文", "张春桥", "江青", "姚文元", "刘晓波", "丹增嘉措", "李洪志", "陈水扁", "黎智英", "艾未未", 
    "李登辉", "李柱铭", "何俊仁", "陈方安生", "达赖", "魏京生",
    # social control
    "维稳", "处突", "社会治安", "反恐怖", "公安工作", "预防犯罪", "收容审查", "治安工作", "大排查", "扫黄打非", 
    "接访", "反邪教",
    # surveillance
    "监控", "监测", "监视", "管控", "监看", "监视系统", "截听", "监控中心", "情报服务", "排查", "监视器",
    "情报搜集", "间谍卫星", "管理网络", "监控器", "监控站", "监控室", "数据采集", 
    # CCP
    "党中央",  "中国共产党", "党支部", "中共中央", "共青团", "共青团中央", "党委", "中央党校",
    # historical events
    "抗日战争", "解放战争", "抗美援朝", "改革开放", "香港回归", "长征", "三大战役", "秋收起义", "南昌起义", "志愿军", 
    "土地改革", "六四", "遵义会议", "广州起义", "百团大战", "文革", "文化大革命", "大跃进", "四人帮",
    # positive figures
    "毛泽东", "江泽民", "胡锦涛", "习近平", "周恩来", "朱镕基", "温家宝", "李克强", "邓小平", "曾庆红", "华国锋", "李鹏", 
    "杨尚昆", "谷牧", "吴邦国", "李岚清", "纪登奎", "乔石","邹家华", "李瑞环", "俞正声", "张高丽", "田纪云", "回良玉", 
    "李源潮", "贾庆林", "姚依林", "张立昌", "尉健行", "姜春云", "李铁映", "王兆国", "罗干", "杨汝岱", "王光英", "彭佩云", 
    "刘云山", "丁关根", "彭真", "胡启立", "曾培炎", "何东昌"]

len(words)

# Baidubaike

# remove punctuation, news agencies, duplicate headlines, numbers & English letters
# test set
news = pd.read_csv('auxiliary/Political_news_titles.csv')
news = news[news['label'].isnull()==False]
punct = ['......', '，', '--', '：',  '丨', '「', '」', '、', '——',  '-',  '_', '...', '·', '。',
        '【', '】', '《', '》', '/', '(图)', '(组图)', '[图]', '‧', '.', ':', ',', '（下）', '（上）', '（图）', '（组图）', ' ',
        '丨', '|', '(', ')', '（', '）', '[', ']', "'", '’', '‘', '“', '”', '%', '?', '!']
for i in range(len(news)):
    text = news.iloc[i,0]
    for p in punct:
      text = text.replace(p, "")
    text = re.sub('[a-zA-Z]','',text)
    text = re.sub("\d+", "", text)
    news.iloc[i,0] = text
news = news[news['label']!=0]
news.drop_duplicates(subset=['sentence'], inplace=True)
news = news.dropna(how='any',axis=0) 
news = news.reset_index(drop=True)
news["label"] = news["label"].astype(int)
news.head()

# training set
train = pd.read_csv('auxiliary/TNEWS_subset.csv')
train = train[train['label'].isnull()==False]
for i in range(len(train)):
    text = train.iloc[i,0]
    for p in punct:
      text = text.replace(p, "")
    text = re.sub('[a-zA-Z]','',text)
    text = re.sub("\d+", "", text)
    train.iloc[i,0] = text
train = train[train['label']!=0]
train.drop_duplicates(subset=['sentence'], inplace=True)
train = train.dropna(how='any',axis=0) 
train = train.reset_index(drop=True)
len(train)

# load Baidubaike word embedding
bd = pd.read_csv('output/2.dropna_baidubaike_clean.csv')
bd.rename(columns={'Unnamed: 0': 'name'}, inplace=True)
bd_wordidx = word2idx(bd)
bd.head()

train_embedding = pd.DataFrame(index=range(0, len(train)), columns=range(0, 300))
test_embedding = pd.DataFrame(index=range(0, len(news)), columns=range(0, 300))

# convert headlines in training set to word embedding
for i in range(0, len(train)):
    text = jieba.lcut(train.iloc[i,0])
    textidx = sen2idx(text, bd_wordidx)
    train_embedding.loc[i] = bd.loc[textidx,'V2':'V301'].mean(axis=0).values
    if i%1000==0:
        print(i)
        
# convert headlines in test set to word embedding
for i in range(0, len(news)):
    text = jieba.lcut(news.iloc[i,0])
    textidx = sen2idx(text, bd_wordidx)
    test_embedding.loc[i] = bd.loc[textidx,'V2':'V301'].mean(axis=0).values
    if i%2500==0:
        print(i)

train_lab = train['label'][train_embedding.isnull().any(axis=1)==False].values
train_embedding = train_embedding.loc[train_embedding.isnull().any(axis=1)==False]
test_lab = news['label'][test_embedding.isnull().any(axis=1)==False].values
news = news.loc[test_embedding.isnull().any(axis=1)==False]
test_embedding = test_embedding.loc[test_embedding.isnull().any(axis=1)==False]
train_bd = preprocessing.scale(train_embedding)
test_bd = preprocessing.scale(test_embedding)

clf = svm.SVC()
clf.fit(train_bd,train_lab)
y_pred=clf.predict(test_bd)
print("Accuracy:",metrics.accuracy_score(test_lab, y_pred))

for i in range(len(news)):
    news.iloc[i, 0] = ' ' + ' '.join(jieba.lcut(news.iloc[i, 0]))+' '

# store prediction
res = pd.DataFrame(index=range(5000), columns=range(20))
category = [range(0, 22), range(22, 37), range(37, 51), range(51, 66), range(66,83), 
            range(83, 95), range(95, 113), range(113, 121), range(121, 140), range(140, 182)]
for n, c in enumerate(category):
    counter = 0
    for w in c:
        counter += 1
        word = ' ' + words[w] + ' '
        id = [i for i in range(len(news)) if word in news.iloc[i, 0]]
        dif = (y_pred[id] - test_lab[id])
        start = 5000 - res.iloc[:,n].isnull().sum()
        end = start + len(dif.tolist())
        res.iloc[start:end, n] = dif.tolist()
        res.iloc[start:end, n+10] = counter
res.to_csv(r'output/svm_bd.csv', header=False, index=False)

clf = GaussianNB()
clf.fit(train_bd,train_lab)
y_pred=clf.predict(test_bd)
print("Accuracy:",metrics.accuracy_score(test_lab, y_pred))

res = pd.DataFrame(index=range(5000), columns=range(20))
category = [range(0, 22), range(22, 37), range(37, 51), range(51, 66), range(66,83), 
            range(83, 95), range(95, 113), range(113, 121), range(121, 140), range(140, 182)]
for n, c in enumerate(category):
    counter = 0
    for w in c:
        counter += 1
        word = ' ' + words[w] + ' '
        id = [i for i in range(len(news)) if word in news.iloc[i, 0]]
        dif = (y_pred[id] - test_lab[id])
        start = 5000 - res.iloc[:,n].isnull().sum()
        end = start + len(dif.tolist())
        res.iloc[start:end, n] = dif.tolist()
        res.iloc[start:end, n+10] = counter
res.to_csv(r'output/nb_bd.csv', header=False, index=False)

# Wikipedia

news = pd.read_csv('auxiliary/Political_news_titles.csv')
news = news[news['label'].isnull()==False]
punct = ['......', '，', '--', '：',  '丨', '「', '」', '、', '——',  '-',  '_', '...', '·', '。',
        '【', '】', '《', '》', '/', '(图)', '(组图)', '[图]', '‧', '.', ':', ',', '（下）', '（上）', '（图）', '（组图）', ' ',
        '丨', '|', '(', ')', '（', '）', '[', ']', "'", '’', '‘', '“', '”', '%', '?', '!']
for i in range(len(news)):
    text = news.iloc[i,0]
    for p in punct:
      text = text.replace(p, "")
    text = re.sub('[a-zA-Z]','',text)
    text = re.sub("\d+", "", text)
    news.iloc[i,0] = text
news = news[news['label']!=0]
news.drop_duplicates(subset=['sentence'], inplace=True)
news = news.dropna(how='any',axis=0) 
news = news.reset_index(drop=True)
news["label"] = news["label"].astype(int)
news.head()

train = pd.read_csv('auxiliary/TNEWS_subset.csv')
train = train[train['label'].isnull()==False]
for i in range(len(train)):
    text = train.iloc[i,0]
    for p in punct:
      text = text.replace(p, "")
    text = re.sub('[a-zA-Z]','',text)
    text = re.sub("\d+", "", text)
    train.iloc[i,0] = text
train = train[train['label']!=0]
train.drop_duplicates(subset=['sentence'], inplace=True)
train = train.dropna(how='any',axis=0) 
train = train.reset_index(drop=True)
len(train)

# load wikipedia word embedding
wk = pd.read_csv('output/2.dropna_wiki_clean.csv')
wk.rename(columns={'Unnamed: 0': 'name'}, inplace=True)
wk_wordidx = word2idx(wk)
wk.head()

train_embedding = pd.DataFrame(index=range(0, len(train)), columns=range(0, 300))
test_embedding = pd.DataFrame(index=range(0, len(news)), columns=range(0, 300))

for i in range(0, len(train)):
    text = jieba.lcut(train.iloc[i,0])
    textidx = sen2idx(text, wk_wordidx)
    train_embedding.loc[i] = wk.loc[textidx,'V2':'V301'].mean(axis=0).values
    if i%1000==0:
        print(i)
        
for i in range(0, len(news)):
    text = jieba.lcut(news.iloc[i,0])
    textidx = sen2idx(text, wk_wordidx)
    test_embedding.loc[i] = wk.loc[textidx,'V2':'V301'].mean(axis=0).values
    if i%2500==0:
        print(i)

train_lab = train['label'][train_embedding.isnull().any(axis=1)==False].values
train_embedding = train_embedding.loc[train_embedding.isnull().any(axis=1)==False]
test_lab = news['label'][test_embedding.isnull().any(axis=1)==False].values
news = news.loc[test_embedding.isnull().any(axis=1)==False]
test_embedding = test_embedding.loc[test_embedding.isnull().any(axis=1)==False]
train_wk = preprocessing.scale(train_embedding)
test_wk = preprocessing.scale(test_embedding)

clf = svm.SVC()
clf.fit(train_wk,train_lab)
y_pred=clf.predict(test_wk)
print("Accuracy:",metrics.accuracy_score(test_lab, y_pred))

for i in range(len(news)):
    news.iloc[i, 0] = ' ' + ' '.join(jieba.lcut(news.iloc[i, 0]))+' '
    
res = pd.DataFrame(index=range(5000), columns=range(20))
category = [range(0, 22), range(22, 37), range(37, 51), range(51, 66), range(66,83), 
            range(83, 95), range(95, 113), range(113, 121), range(121, 140), range(140, 182)]
for n, c in enumerate(category):
    counter = 0
    for w in c:
        counter += 1
        word = ' ' + words[w] + ' '
        id = [i for i in range(len(news)) if word in news.iloc[i, 0]]
        dif = (y_pred[id] - test_lab[id])
        start = 5000 - res.iloc[:,n].isnull().sum()
        end = start + len(dif.tolist())
        res.iloc[start:end, n] = dif.tolist()
        res.iloc[start:end, n+10] = counter
res.to_csv(r'output/svm_wk.csv', header=False, index=False)

clf = GaussianNB()
clf.fit(train_wk,train_lab)
y_pred=clf.predict(test_wk)
print("Accuracy:",metrics.accuracy_score(test_lab, y_pred))

res = pd.DataFrame(index=range(5000), columns=range(20))
category = [range(0, 22), range(22, 37), range(37, 51), range(51, 66), range(66,83), 
            range(83, 95), range(95, 113), range(113, 121), range(121, 140), range(140, 182)]
for n, c in enumerate(category):
    counter = 0
    for w in c:
        counter += 1
        word = ' ' + words[w] + ' '
        id = [i for i in range(len(news)) if word in news.iloc[i, 0]]
        dif = (y_pred[id] - test_lab[id])
        start = 5000 - res.iloc[:,n].isnull().sum()
        end = start + len(dif.tolist())
        res.iloc[start:end, n] = dif.tolist()
        res.iloc[start:end, n+10] = counter
res.to_csv(r'output/nb_wk.csv', header=False, index=False)

# People's Daily

news = pd.read_csv('auxiliary/Political_news_titles.csv')
news = news[news['label'].isnull()==False]
punct = ['......', '，', '--', '：',  '丨', '「', '」', '、', '——',  '-',  '_', '...', '·', '。',
        '【', '】', '《', '》', '/', '(图)', '(组图)', '[图]', '‧', '.', ':', ',', '（下）', '（上）', '（图）', '（组图）', ' ',
        '丨', '|', '(', ')', '（', '）', '[', ']', "'", '’', '‘', '“', '”', '%', '?', '!']
for i in range(len(news)):
    text = news.iloc[i,0]
    for p in punct:
      text = text.replace(p, "")
    text = re.sub('[a-zA-Z]','',text)
    text = re.sub("\d+", "", text)
    news.iloc[i,0] = text
news = news[news['label']!=0]
news.drop_duplicates(subset=['sentence'], inplace=True)
news = news.dropna(how='any',axis=0) 
news = news.reset_index(drop=True)
news["label"] = news["label"].astype(int)

train = pd.read_csv('auxiliary/TNEWS_subset.csv')
train = train[train['label'].isnull()==False]
for i in range(len(train)):
    text = train.iloc[i,0]
    for p in punct:
      text = text.replace(p, "")
    text = re.sub('[a-zA-Z]','',text)
    text = re.sub("\d+", "", text)
    train.iloc[i,0] = text
train = train[train['label']!=0]
train.drop_duplicates(subset=['sentence'], inplace=True)
train = train.dropna(how='any',axis=0) 
train = train.reset_index(drop=True)
len(train)

# load wikipedia word embedding
rm = pd.read_csv('output/2.dropna_renmin_clean.csv')
rm.rename(columns={'Unnamed: 0': 'name'}, inplace=True)
rm_wordidx = word2idx(rm)
rm.head()

train_embedding = pd.DataFrame(index=range(0, len(train)), columns=range(0, 300))
test_embedding = pd.DataFrame(index=range(0, len(news)), columns=range(0, 300))

for i in range(0, len(train)):
    text = jieba.lcut(train.iloc[i,0])
    textidx = sen2idx(text, rm_wordidx)
    train_embedding.loc[i] = rm.loc[textidx,'V2':'V301'].mean(axis=0).values
    if i%1000==0:
        print(i)
        
for i in range(0, len(news)):
    text = jieba.lcut(news.iloc[i,0])
    textidx = sen2idx(text, rm_wordidx)
    test_embedding.loc[i] = rm.loc[textidx,'V2':'V301'].mean(axis=0).values
    if i%2500==0:
        print(i)

train_lab = train['label'][train_embedding.isnull().any(axis=1)==False].values
train_embedding = train_embedding.loc[train_embedding.isnull().any(axis=1)==False]
test_lab = news['label'][test_embedding.isnull().any(axis=1)==False].values
news = news.loc[test_embedding.isnull().any(axis=1)==False]
test_embedding = test_embedding.loc[test_embedding.isnull().any(axis=1)==False]
train_rm = preprocessing.scale(train_embedding)
test_rm = preprocessing.scale(test_embedding)

clf = svm.SVC()
clf.fit(train_rm,train_lab)
y_pred=clf.predict(test_rm)
print("Accuracy:",metrics.accuracy_score(test_lab, y_pred))

for i in range(len(news)):
    news.iloc[i, 0] = ' ' + ' '.join(jieba.lcut(news.iloc[i, 0]))+' '
    
res = pd.DataFrame(index=range(5000), columns=range(20))
category = [range(0, 22), range(22, 37), range(37, 51), range(51, 66), range(66,83), 
            range(83, 95), range(95, 113), range(113, 121), range(121, 140), range(140, 182)]
for n, c in enumerate(category):
    counter = 0
    for w in c:
        counter += 1
        word = ' ' + words[w] + ' '
        id = [i for i in range(len(news)) if word in news.iloc[i, 0]]
        dif = (y_pred[id] - test_lab[id])
        start = 5000 - res.iloc[:,n].isnull().sum()
        end = start + len(dif.tolist())
        res.iloc[start:end, n] = dif.tolist()
        res.iloc[start:end, n+10] = counter
res.to_csv(r'output/svm_rm.csv', header=False, index=False)

clf = GaussianNB()
clf.fit(train_rm,train_lab)
y_pred=clf.predict(test_rm)
print("Accuracy:",metrics.accuracy_score(test_lab, y_pred))

res = pd.DataFrame(index=range(5000), columns=range(20))
category = [range(0, 22), range(22, 37), range(37, 51), range(51, 66), range(66,83), 
            range(83, 95), range(95, 113), range(113, 121), range(121, 140), range(140, 182)]
for n, c in enumerate(category):
    counter = 0
    for w in c:
        counter += 1
        word = ' ' + words[w] + ' '
        id = [i for i in range(len(news)) if word in news.iloc[i, 0]]
        dif = (y_pred[id] - test_lab[id])
        start = 5000 - res.iloc[:,n].isnull().sum()
        end = start + len(dif.tolist())
        res.iloc[start:end, n] = dif.tolist()
        res.iloc[start:end, n+10] = counter
res.to_csv(r'output/nb_rm.csv', header=False, index=False)
