==> 4.Preprocess_TextCNN.ipynb <==
import numpy as np
import pandas as pd
import pickle
import os
import re
import jieba
import random
import math

def word2idx(embedding):
    wordidx = {}
    for i in range(len(embedding)):
        word=embedding['name'][i]
        wordidx[word] = len(wordidx)
    return(wordidx)           
            
def sen2idx(sen, wordidx):
    idxword = []
    for s in sen:
        try:
            idx = wordidx[s]
            idxword.append(idx)
        except:
            continue
    return(idxword)

os.chdir(r'/Replication_TrainingDatasetCensorship') # set working directory

wk = pd.read_csv('output/2.dropna_wiki_clean.csv')
wk.rename(columns={'Unnamed: 0': 'name'}, inplace=True)
wk = wk.dropna()
wk.drop_duplicates(subset=['name'], inplace=True)
wk = wk.reset_index(drop=True)
wk.head()

wknp = wk.loc[:,'V2':'V301'].values
np.savez('TextCNN/data/wknp.npz', wknp)

wk_wordidx = word2idx(wk)
with open('TextCNN/data/vocab_wk.pkl', 'wb') as f:
    pickle.dump(wk_wordidx, f)

bd = pd.read_csv('output/2.dropna_baidubaike_clean.csv')
bd.rename(columns={'Unnamed: 0': 'name'}, inplace=True)
bd = bd.dropna()
bd.drop_duplicates(subset=['name'], inplace=True)
bd = bd.reset_index(drop=True)
bdnp = bd.loc[:,'V2':'V301'].values
np.savez('TextCNN/data/bdnp.npz', bdnp)
bd_wordidx = word2idx(bd)
with open('TextCNN/data/vocab_bd.pkl', 'wb') as f:
    pickle.dump(bd_wordidx, f)

rm = pd.read_csv('output/2.dropna_renmin_clean.csv')
rm.rename(columns={'Unnamed: 0': 'name'}, inplace=True)
rm = rm.dropna()
rm.drop_duplicates(subset=['name'], inplace=True)
rm = rm.reset_index(drop=True)
rmnp = rm.loc[:,'V2':'V301'].values
np.savez('TextCNN/data/rmnp.npz', rmnp)
rm_wordidx = word2idx(rm)
with open('TextCNN/data/vocab_rm.pkl', 'wb') as f:
    pickle.dump(rm_wordidx, f)

news = pd.read_csv('auxiliary/Political_news_titles.csv')
news = news[news['label'].isnull()==False]
punct = ['......', '，', '--', '：',  '丨', '「', '」', '、', '——',  '-',  '_', '...', '·', '。',
        '【', '】', '《', '》', '/', '(图)', '(组图)', '[图]', '‧', '.', ':', ',', '（下）', '（上）', '（图）', '（组图）', ' ',
        '丨', '|', '(', ')', '（', '）', '[', ']', "'", '’', '‘', '“', '”', '%', '?', '!']
for i in range(len(news)):
    text = news.iloc[i,0]
    for p in punct:
      text = text.replace(p, "")
    text = re.sub('[a-zA-Z]','',text)
    text = re.sub("\d+", "", text)
    news.iloc[i,0] = text
news = news[news['label']!=0]
news.drop_duplicates(subset=['sentence'], inplace=True)
news = news.dropna(how='any',axis=0) 
news = news.reset_index(drop=True)
news["label"] = news["label"].astype(int)

for i in range(len(news)):
    if news.iloc[i, 1]==-1:
        news.iloc[i, 1] = 0
news.head()

train = pd.read_csv('auxiliary/TNEWS_subset.csv')
train = train[train['label'].isnull()==False]
for i in range(len(train)):
    text = train.iloc[i,0]
    for p in punct:
      text = text.replace(p, "")
    text = re.sub('[a-zA-Z]','',text)
    text = re.sub("\d+", "", text)
    train.iloc[i,0] = text
train = train[train['label']!=0]
train.drop_duplicates(subset=['sentence'], inplace=True)
train = train.dropna(how='any',axis=0) 
train = train.reset_index(drop=True)

for i in range(len(train)):
    if train.iloc[i, 1]==-1:
        train.iloc[i, 1] = 0
train.head()

jieba.load_userdict("auxiliary/Dict.txt")

for i in range(len(news)):
    c = jieba.lcut(news.iloc[i, 0])
    news.iloc[i, 0] = ' '.join(c)
news.head()

for i in range(len(train)):
    c = jieba.lcut(train.iloc[i, 0])
    train.iloc[i, 0] = ' '.join(c)
train.head()

random.seed(92092)
dev_idx = random.sample(range(len(train)), math.floor(len(train)*0.2))
len(dev_idx)

train_idx = [j for j in range(len(train)) if j not in dev_idx] 
len(train_idx)

trainset = train.iloc[train_idx]
dev = train.iloc[dev_idx]

news = news.to_dict('records')
train = trainset.to_dict('records')
dev = dev.to_dict('records')

with open('TextCNN/data/train.txt', 'w') as fp:
    for t in train:
        fp.write(t['sentence'] + '__' + str(t['label']) + '\n')

with open('TextCNN/data/dev.txt', 'w') as fp:
    for t in dev:
        fp.write(t['sentence'] + '__' + str(t['label']) + '\n')

with open('TextCNN/data/test.txt', 'w') as fp:
    for t in news:
        fp.write(t['sentence'] + '__' + str(t['label']) + '\n')
