==> 4.TextCNN.ipynb <==
gpu_info = !nvidia-smi
gpu_info = '\n'.join(gpu_info)
if gpu_info.find('failed') >= 0:
  print('Select the Runtime > "Change runtime type" menu to enable a GPU accelerator, ')
  print('and then re-execute this cell.')
else:
  print(gpu_info)

pip install tensorboardX

import sys
import numpy as np
sys.path.append('/content/drive/My Drive/textcnn/')

%run '/content/drive/My Drive/textcnn/run1.py' --model TextCNN --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/wk1.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/run2.py' --model TextCNN --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/wk2.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/run3.py' --model TextCNN --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/wk3.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/run4.py' --model TextCNN --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/wk4.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/run5.py' --model TextCNN --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/wk5.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/run6.py' --model TextCNN --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/wk6.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/run7.py' --model TextCNN --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/wk7.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/run8.py' --model TextCNN --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/wk8.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/run9.py' --model TextCNN --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/wk9.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/run10.py' --model TextCNN --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/wk10.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/runbd1.py' --model TextCNNBD --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/bd1.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/runbd2.py' --model TextCNNBD --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/bd2.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/runbd3.py' --model TextCNNBD --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/bd3.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/runbd4.py' --model TextCNNBD --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/bd4.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/runbd5.py' --model TextCNNBD --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/bd5.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/runbd6.py' --model TextCNNBD --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/bd6.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/runbd7.py' --model TextCNNBD --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/bd7.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/runbd8.py' --model TextCNNBD --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/bd8.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/runbd9.py' --model TextCNNBD --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/bd9.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/runbd10.py' --model TextCNNBD --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/bd10.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/runrm1.py' --model TextCNNRM --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/rm1.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/runrm2.py' --model TextCNNRM --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/rm2.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/runrm3.py' --model TextCNNRM --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/rm3.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/runrm4.py' --model TextCNNRM --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/rm4.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/runrm5.py' --model TextCNNRM --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/rm5.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/runrm6.py' --model TextCNNRM --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/rm6.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/runrm7.py' --model TextCNNRM --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/rm7.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/runrm8.py' --model TextCNNRM --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/rm8.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/runrm9.py' --model TextCNNRM --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/rm9.txt", predict_all, delimiter=",", fmt="%d")

%run '/content/drive/My Drive/textcnn/runrm10.py' --model TextCNNRM --word True
predict_all[predict_all==0] = -1
np.savetxt("/content/drive/MyDrive/textcnn/output/rm10.txt", predict_all, delimiter=",", fmt="%d")
